Renamed Card.js to VehicleCard.js to avoid interfering with materialize 'Card' and restyled inventory cards.
